# Conflicts:
#	PRIS.Web/Startup.cs

+++ PRIS.Core.Library/Entities/ConversationResult.cs
@@ -1,10 +1,10 @@
-﻿using System.Collections.Generic;
+﻿    using System.Collections.Generic;
 
 namespace PRIS.Core.Library.Entities
 {
     public class ConversationResult : EntityBase
     {
-        public int? Grade { get; set; }
+        public double? Grade { get; set; }
         public string Comment { get; set; }
 
         public List<Student> Students { get; set; }
